Add a new component to track spells from DisplayHub
