+++ src/components/App/App.js
@@ -1,10 +1,9 @@
 import './App.css';
 import { Route, Routes } from "react-router-dom"
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 import Header from '../Header/Header';
-import cleanSpells from '../../util/cleaning';
 import Homepage from '../Homepage/Homepage.js'
-import Display from '../Display/Display';
+import DisplayHub from '../DisplayHub/DisplayHub';
 import DecisionPage from "../DecisionPage/decisionPage";
 
 
@@ -23,7 +22,7 @@ function App() {
              <main className='App'>
                  <Routes>
                      <Route path="/" element={<Homepage handleName={handleName}/>} />
-                     <Route path="/display" element={<Display />} />
+                     <Route path="/display" element={<DisplayHub/>} />
                      <Route path="/decision" element={<DecisionPage user={personName} />} />
                      {/* <Route path="/" element={<Homepage />} /> */}
                      {/* <Route path="/" element={<Homepage />} /> */}
