+++ src/components/DisplayHub/DisplayHub.css
@@ -1,0 +1,7 @@
+.display-area{
+  border:4px solid black; 
+  margin-top: 5.5rem;
+  height: 100vh;
+}
+
+
