+++ src/components/SpellsArea/SpellsArea.css
@@ -1,18 +1,15 @@
-.display-area{
-  border:4px solid black; 
-  margin-top: 5.5rem;
-}
-
 .spell-card-wrapper{
   border:1px solid black;
-  height:40%;
-  width: 30%; 
-  margin: 1rem 1rem;
+  height:30%;
+  width: 20%; 
+  margin: 1rem 2rem;
   box-shadow: 2px 2px 5px 0 rgba(0, 0, 0, .2)
 }
-
-.spells-area {
+.spells-displayed {
   display: flex;
   flex-direction: row;
   flex-wrap: wrap;
+  width: 100vw;
+  overflow-y: auto;
+  height:100vh;
 }
